requirements.txt
PyByteBuffer
Arc4
python-dotenv
lxml
python-telegram-bot